Eye sprites - reorganize wip #101
