+++ src/sprites/eyes/circle-chain-dense.ts
