+++ src/sprites/eyes/pointy-chain.ts
@@ -1,0 +1,27 @@
+import { EyeSprites } from "../EyeSprites";
+
+const definitions = `<path id="eye" x="0" y="0" d="M676.9,16.9l-9.5-16.5h-52l-8.5,14.7L581.3,0.4l-33.5,19.3L514.3,0.4l-33.5,19.3L447.3,0.4
+l-33.5,19.3L380.3,0.4l-33.5,19.3L313.3,0.4l-33.5,19.3L246.3,0.4l-33.5,19.3L179.3,0.4l-33.5,19.3L112.3,0.4L86.7,15.2L78.2,0.4
+h-52l-9.5,16.5L0.3,26.4v52L15,86.9L0.3,112.4l19.3,33.5L0.3,179.4l19.3,33.5L0.3,246.4l19.3,33.5L0.3,313.5L19.6,347L0.3,380.5
+L19.6,414L0.3,447.5L19.6,481L0.3,514.5L19.6,548L0.3,581.5L15,607l-14.7,8.5v52l16.5,9.5l9.5,16.5h52l8.5-14.7l25.5,14.7
+l33.5-19.3l33.5,19.3l33.5-19.3l33.5,19.3l33.5-19.3l33.5,19.3l33.5-19.3l33.5,19.3l33.5-19.3l33.5,19.3l33.5-19.3l33.5,19.3
+l33.5-19.3l33.5,19.3l25.5-14.7l8.5,14.7h52l9.5-16.5l16.5-9.5v-52l-14.7-8.5l14.7-25.5L674,548l19.3-33.5L674,481l19.3-33.5
+L674,414l19.3-33.5L674,347l19.3-33.5L674,279.9l19.3-33.5L674,212.9l19.3-33.5L674,145.9l19.3-33.5l-14.7-25.5l14.7-8.5v-52
+L676.9,16.9z M547.8,608.9l-33.5-19.3l-33.5,19.3l-33.5-19.3l-33.5,19.3l-33.5-19.3l-33.5,19.3l-33.5-19.3l-33.5,19.3l-33.5-19.3
+l-33.5,19.3l-33.5-19.3l-33.5,19.3l-33.5-19.3l-19.1,11l11-19.1L84.9,548l19.3-33.5L84.9,481l19.3-33.5L84.9,414l19.3-33.5
+L84.9,347l19.3-33.5L84.9,280l19.3-33.5L84.9,213l19.3-33.5L84.9,146l19.3-33.5l-11-19.1l19.1,11L145.8,85l33.5,19.3L212.8,85
+l33.5,19.3L279.8,85l33.5,19.3L346.8,85l33.5,19.3L413.8,85l33.5,19.3L480.8,85l33.5,19.3L547.8,85l33.5,19.3l19.1-11l-11,19.1
+l19.3,33.5l-19.3,33.5l19.3,33.5l-19.3,33.5l19.3,33.5l-19.3,33.5l19.3,33.5l-19.3,33.5l19.3,33.5l-19.3,33.5l19.3,33.5l-19.3,33.5
+l19.3,33.5l-19.3,33.5l11,19.1l-19.1-11L547.8,608.9z">`;
+
+const useIds = {
+    "topleft": "eye",
+    "topright": "eye",
+    "bottomleft": "eye",
+};
+
+export const sprites = new EyeSprites(
+    `Pointy Chain`,
+    definitions,
+    useIds
+);
