+++ src/sprites/eyes/pointy-skewed.ts
@@ -6,7 +6,7 @@ const definitions = `
     <path id="eye-tr" d="M549.7,700.4c46.5,0,90.7-48,97.7-92.4L699.9,0.4L83.1,39.9C39.7,46.4,5.5,87.5,5.4,136.8
     L0.1,700.4H549.7z M100.1,600.4l3.8-402.4c0-35.2,24.5-64.6,55.5-69.3L600,100.5l-37.5,433.9c-5,31.8-36.6,66.1-69.8,66.1
     L100.1,600.4L100.1,600.4z"/>
-    <path id="eye-bl"  d="$M150.6,0.1c-46.5,0-90.7,48-97.7,92.4L0.3,700.1l616.8-39.5c43.5-6.5,77.7-47.6,77.7-96.9
+    <path id="eye-bl"  d="M150.6,0.1c-46.5,0-90.7,48-97.7,92.4L0.3,700.1l616.8-39.5c43.5-6.5,77.7-47.6,77.7-96.9
     l5.3-563.5L150.6,0.1L150.6,0.1z M600.2,100.1l-3.8,402.4c0,35.2-24.5,64.6-55.5,69.3L100.3,600l37.5-433.9
     c5-31.8,36.6-66.1,69.8-66.1L600.2,100.1L600.2,100.1z"/>
         `;
@@ -18,7 +18,7 @@ const useIds: IEyePosObj = {
 };
 
 export const sprites = new EyeSprites(
-    `Style 3`,
+    `Pointy skewed`,
     definitions,
     useIds
 );
